requirements.txt
Django==1.7.6
django-authtools==1.2.0
django-braces==1.8.1
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.2
django-extensions==1.5.5
django-mssql==1.6.2
django-organizations==0.5.3
django-simple-history==1.6.3
django-sqlserver==1.7
pypiwin32==219
pytz==2015.4
six==1.9.0
sqlparse==0.1.16
